readme please :)
In a rush, I managed to export an odd sample with the final track and did not notice until much later. I found it a little bit inappropriate (spooky honestly), and though it would be best to exchange it :) Thank you for your patience with the matter, as I've emailed you about it, too.